put device in recovery mode _after_ extracting the root filesystem
Since we can grab tss records in normal mode and these depend on
nonce values that are valid only for the first reboot, we should not
extract the filesystem after putting the device into recovery mode.
If the filesystem extraction takes longer (e.g. on a network filesystem)
the recovery mode device might reboot while the file system is still
extracting. In that case the TSS record that has been grabbed in normal
mode would be invalid.