Select build identity based on hardware model instead of first or last identity in manifest
While this might have worked in the past it didn't work anymore with the
iPhone 6S (plus) since it comes with two different CPUs (Samsung & TSMC A9).
Therefore the BuildManifest.plist has multiple build identities and for a
successful restore the correct build identity has to be selected for the
actual hardware model of the corresponding device.