restore: Make sure newer Mav16 baseband firmware image files are recognized
The iPhone 7/7+ models with Mav16 baseband failed to restore as the baseband
firmware files are now called RestoreSBL1-HastTableDigest and
SBL1-HashTableDigest. The code only stripped off the '-Blob' part from the
baseband tss response so it couldn't match the firmware filenames properly.
This commit fixes the issue by only keeping the first part of the string
(everything before the first occurance of '-').