Change check_hardware_model() to get_irecv_device()
The check_hardware_model() function has a misleading name. It returns
a string with the hardware model but this is not used. Instead, this
function is only called for its side effect to initialise an irecv
device in the passed client struct which it creates from the hardware
model name returned by mode specific implementations. But these mode
specific implementations already create an irecv device to get the
hardware model name so instead of going through this unnecessary
complication just return the irecv device directly and rename the
function accordingly to make this clear. (This may also prevent
leaking an irecv device in the mode specific functions.)