restore: Fix NORData request by sending new components (fixes wireless charging on newer devices)
The NorImageData response didn't include the WCHFirmwareUpdater for newer
devices, resulting in wireless charging capability to be broken on iPhone XS,
iPhone XR, and newer. Upon further inspection it turned out that the selection
for the images to send was only based on `IsFirmwarePayload` property.
However, there are additional components with other properties, as the
WCHFirmwareUpdater one, that don't have the `IsFirmwarePayload` property but
instead `IsSecondaryFirmwarePayload` and `IsLoadedByiBoot` which seem to be
the rule for including these images.