restore: Fixed a problem that nobody even knew existed
At least this has been going unnoticed until recently. For quite some time
we have been sending NorImageData as array to the device, but it turned out
that this was only expected for iOS < 7.0 and from then on it was supposed
to be a dictionary with the components.

Now we should correctly handle it.