mmap a component file instead of malloc and fread
Currently, we read an entire component file into memory allocated by
malloc. This always requires the same amount of memory as the file
size and it can be a few or more gigabytes with a certain IPSW.
Instead, just map the files via mmap(2) to avoid having such a copy in
user process's memory. The file is mapped as a private copy-on-write
mapping so that we can update the mapped data while not carrying
through the change to the underlying file.

With this change, a component now needs to be extracted and *free*d
with ipsw_extract_component() and ipsw_free_component() respectively
otherwise the component may not be properly unmapped/freed.

Obviously, this only works with a fully extracted IPSW. Also, win32
build still uses malloc and fread as it doesn't have mmap.

Signed-off-by: Munehisa Kamata <kamatam@amazon.com>