recovery: set bRequest to 1 when sending bootx command
In macOS 13 beta 8 or newer release, bootx seems to fail if bRequest
is 0 in the control transfer setup. Then, the device fails to enter
restore mode. Seems like something has changed in iBEC since beta 8 and
Apple Configurator 2 has set it to 1, so do the same thing.

While this could be applied for all *OS variants, it's limited to macOS
for now just to be safe.

Signed-off-by: Munehisa Kamata <kamatam@amazon.com>