Print message about USB port for macOS restores in case Port DFU fails
Depending on the model, a specific USB Port has to be used for a restore.
When using a different one, Port DFU fails to switch to DFU mode. We use
this fact to print a message with an official link that shows the correct
USB port to use for the restore.